Adding integration tests
This commit adds integration tests along with an ASP.NET Core web site
that has multiple types of authentication enabled. This helps to
validate that the implementation is correct when executed against an
actual server.

As an additional aside, the unit test project (PassedBall.UnitTests)
has been updated to a .NET Core 2.x project as well. The main
PassedBall project remains a multitargeted project, generating
assemblies for .NET Framework 4.5, 4.6, 4.7, and .NET Standard 2.0.